+++ packages/ui/src/lib/WalletMultiButton.svelte
@@ -3,6 +3,7 @@
     import WalletButton from './WalletButton.svelte';
     import WalletConnectButton from './WalletConnectButton.svelte';
     import WalletModal from './WalletModal.svelte';
+    import { clickOutside } from './clickOutside';
     import './styles.css';
 
     export let maxNumberOfWallets = 3;
@@ -47,37 +48,6 @@
         closeDropdown();
         await disconnect();
     }
-
-    interface CallbackType {
-        (arg?: string): void;
-    }
-
-    function clickOutside(
-        node: HTMLElement,
-        callbackFunction: CallbackType,
-    ): unknown {
-        function onClick(event: MouseEvent) {
-            if (
-                node &&
-                event.target instanceof Node &&
-                !node.contains(event.target) &&
-                !event.defaultPrevented
-            ) {
-                callbackFunction();
-            }
-        }
-
-        document.body.addEventListener('click', onClick, true);
-
-        return {
-            update(newCallbackFunction: CallbackType) {
-                callbackFunction = newCallbackFunction;
-            },
-            destroy() {
-                document.body.removeEventListener('click', onClick, true);
-            },
-        };
-    }
 </script>
 
 {#if !wallet}
@@ -108,26 +78,20 @@
                     }
                 }}
             >
-                <li
-                    on:click={copyAddress}
-                    class="wallet-adapter-dropdown-list-item"
-                    role="menuitem"
-                >
-                    {copied ? 'Copied' : 'Copy address'}
+                <li role="menuitem">
+                    <button class="wallet-adapter-dropdown-list-item" on:click={copyAddress}>
+                        {copied ? 'Copied' : 'Copy address'}
+                    </button>
                 </li>
-                <li
-                    on:click={openModal}
-                    class="wallet-adapter-dropdown-list-item"
-                    role="menuitem"
-                >
-                    Connect a different wallet
+                <li role="menuitem">
+                    <button class="wallet-adapter-dropdown-list-item" on:click={openModal}>
+                        Connect a different wallet
+                    </button>
                 </li>
-                <li
-                    on:click={disconnectWallet}
-                    class="wallet-adapter-dropdown-list-item"
-                    role="menuitem"
-                >
-                    Disconnect
+                <li role="menuitem">
+                    <button class="wallet-adapter-dropdown-list-item" on:click={disconnectWallet}>
+                        Disconnect
+                    </button>
                 </li>
             </ul>
         {/if}
